// apps/web/lib/api/schemas/graphEdge.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Gnosis API
 * Real-time semantic mind-mapping engine
 * OpenAPI spec version: 1.0.0
 */
import type { GraphEdgeLabel } from './graphEdgeLabel';

export interface GraphEdge {
  /**
   * @minLength 1
   * @maxLength 100
   */
  source: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  target: string;
  /** Relationship type */
  label?: GraphEdgeLabel;
}